+++ runtime/objc/MOBStateMachine/MOBStateMachineTests/testCaseDefinitions/src-gen/PasswordEntryStateMachine.m
@@ -235,7 +235,7 @@
 				[self switchTransitionWithID: PasswordEntry_Transition_EMPTY_PARTIALLYFILLED];
 			}
 			else {
-				NSLog(@"MOBStateMachine::No guard permits transition in state: %@ for event: %d.", [self.currentState name], event);
+				NSLog(@"MOBStateMachine::No guard permits transition in state: %@ for event: %lu.", [self.currentState name], (unsigned long)event);
 			}
 			break;
 		}
@@ -249,14 +249,14 @@
 				[self switchTransitionWithID: PasswordEntry_Transition_PARTIALLYFILLED_PARTIALLYFILLED_DIGITENTERED];
 			}
 			else {
-				NSLog(@"MOBStateMachine::No guard permits transition in state: %@ for event: %d.", [self.currentState name], event);
+				NSLog(@"MOBStateMachine::No guard permits transition in state: %@ for event: %lu.", [self.currentState name], (unsigned long)event);
 			}
 			break;
 		}
 
 		default:
 		{
-			NSLog(@"MOBStateMachine::Warning: state: %@ not designated to react to event: %d.", [self.currentState name], event);
+			NSLog(@"MOBStateMachine::Warning: state: %@ not designated to react to event: %lu.", [self.currentState name], (unsigned long)event);
 			break;
 		}
 		}
@@ -275,14 +275,14 @@
 				[self switchTransitionWithID: PasswordEntry_Transition_COMPLETELYFILLED_EMPTY];
 			}
 			else {
-				NSLog(@"MOBStateMachine::No guard permits transition in state: %@ for event: %d.", [self.currentState name], event);
+				NSLog(@"MOBStateMachine::No guard permits transition in state: %@ for event: %lu.", [self.currentState name], (unsigned long)event);
 			}
 			break;
 		}
 
 		default:
 		{
-			NSLog(@"MOBStateMachine::Warning: state: %@ not designated to react to event: %d.", [self.currentState name], event);
+			NSLog(@"MOBStateMachine::Warning: state: %@ not designated to react to event: %lu.", [self.currentState name], (unsigned long)event);
 			break;
 		}
 		}
@@ -298,14 +298,14 @@
 				[self switchTransitionWithID: PasswordEntry_Transition_COMPLETELYFILLED_FINAL0];
 			}
 			else {
-				NSLog(@"MOBStateMachine::No guard permits transition in state: %@ for event: %d.", [self.currentState name], event);
+				NSLog(@"MOBStateMachine::No guard permits transition in state: %@ for event: %lu.", [self.currentState name], (unsigned long)event);
 			}
 			break;
 		}
 
 		default:
 		{
-			NSLog(@"MOBStateMachine::Warning: state: %@ not designated to react to event: %d.", [self.currentState name], event);
+			NSLog(@"MOBStateMachine::Warning: state: %@ not designated to react to event: %lu.", [self.currentState name], (unsigned long)event);
 			break;
 		}
 		}
@@ -324,14 +324,14 @@
 				[self switchTransitionWithID: PasswordEntry_Transition_PARTIALLYFILLED_PARTIALLYFILLED_DIGITDELETED];
 			}
 			else {
-				NSLog(@"MOBStateMachine::No guard permits transition in state: %@ for event: %d.", [self.currentState name], event);
+				NSLog(@"MOBStateMachine::No guard permits transition in state: %@ for event: %lu.", [self.currentState name], (unsigned long)event);
 			}
 			break;
 		}
 
 		default:
 		{
-			NSLog(@"MOBStateMachine::Warning: state: %@ not designated to react to event: %d.", [self.currentState name], event);
+			NSLog(@"MOBStateMachine::Warning: state: %@ not designated to react to event: %lu.", [self.currentState name], (unsigned long)event);
 			break;
 		}
 		}
@@ -340,7 +340,7 @@
 
 	default:
 	{
-		NSLog(@"MOBStateMachine::Error: Cannot handle event: %d", event);
+		NSLog(@"MOBStateMachine::Error: Cannot handle event: %lu", (unsigned long)event);
 		[self.delegate handleStateMachineError: MOBStateMachineError_UnknownEvent];
 	}
 	}
